move two processes
